Support creating FuncTypes from vectors of kinds
This introduces the `FuncType` constructor:

```
(.->.) :: V.Vector Kind -> V.Vector Kind -> FuncType
```

which, as it type indicates, constructs a `FuncType` from a vectos of
kinds for the parameters and result of the function respectively.

This also finishes the `Show` instance for `FuncType` to use this
constructor. For example, printing the `moduleExports` of the `gcd`
example shows:

```
[newExportType "gcd" (ExternFuncType ([KindI32,KindI32] .->. [KindI32]))]
```