fix!: support new wasmtime(v29) (#48)
Is it wasmtime-hs can support wasmtime version 29.0.1.

This pull request includes several changes to the `flake.nix`
configuration and updates to the Wasmtime Haskell bindings. The most
important changes include the removal of the Wasmtime input from the Nix
flake, updates to the Wasmtime configuration functions, and the addition
of a buffering specification in a test file.

### Nix Configuration Updates:
* Removed the Wasmtime input from the `flake.nix` file and added an
overlay to build Wasmtime with C API support.
[[1]](diffhunk://#diff-206b9ce276ab5971a2489d75eb1b12999d4bf3843b7988cbe8d687cfde61dea0L7-R30)
[[2]](diffhunk://#diff-206b9ce276ab5971a2489d75eb1b12999d4bf3843b7988cbe8d687cfde61dea0L25-R50)

### Wasmtime Haskell Bindings Updates:
* Replaced removed Wasmtime configuration functions with new ones in
`lib/Bindings/Wasmtime/Config.hsc` and `lib/Wasmtime.hs`.
[[1]](diffhunk://#diff-e98fea86c4690e7b94832c5dfe0184f51fb15dfc4d38eb1c6f5c340ffd083c79L53-R55)
[[2]](diffhunk://#diff-d49ac0017ec35aa1a96fc2a1f36b5cef18166aa8e7ceaa7b1239df2adc7fe0a4L48-R50)
[[3]](diffhunk://#diff-d49ac0017ec35aa1a96fc2a1f36b5cef18166aa8e7ceaa7b1239df2adc7fe0a4L472-R481)

### Test File Update:
* Added a buffering specification in `test/memory.hs` to ensure the test
runs correctly in the `nix build`.